tidy up config files (#65)
* tidy up config files

* tidy up config files 2

* tidy up config files 3

* tidy up config files 4
